Move internals from commands.hh to separate header
