+++ src/phantasm-hardware-interface/vulkan/cmd_buf_translation.cc
@@ -1,6 +1,7 @@
 #include "cmd_buf_translation.hh"
 
 #include <phantasm-hardware-interface/detail/byte_util.hh>
+#include <phantasm-hardware-interface/detail/command_reading.hh>
 #include <phantasm-hardware-interface/detail/incomplete_state_cache.hh>
 
 #include "Swapchain.hh"
@@ -452,9 +453,9 @@ void phi::vk::command_list_translator::execute(const phi::cmd::update_top_level&
 void phi::vk::command_list_translator::execute(const cmd::dispatch_rays& dispatch_rays) {}
 
 void phi::vk::command_list_translator::bind_shader_arguments(phi::handle::pipeline_state pso,
-                                                                     const std::byte* root_consts,
-                                                                     cc::span<const phi::shader_argument> shader_args,
-                                                                     VkPipelineBindPoint bind_point)
+                                                             const std::byte* root_consts,
+                                                             cc::span<const phi::shader_argument> shader_args,
+                                                             VkPipelineBindPoint bind_point)
 {
     auto const& pso_node = _globals.pool_pipeline_states->get(pso);
     pipeline_layout const& pipeline_layout = *pso_node.associated_pipeline_layout;
